configured to make Client2 have data from json generated file and Client1 have data from rabbitmq in the collection

+++ mongodb/mongodb_send_data_from_json_file.py
@@ -5,23 +5,26 @@ import json
 project_fullpath = "/media/stephane/DATA/ESILV/A5/Dev Apps et Web services pour l'IOT/TP/Fichiers TP/Dev app & web service for IOT - Fil rouge"
 with open(project_fullpath+'/generated_data.txt') as json_file:
     data = json.load(json_file)
+    
+    clients_to_ignore = ["Client1"]
 
     databases = {}
     connection = pymongo.MongoClient("mongodb://localhost:27017/")
 
     for client in data:
-        databases[client] = {}
-        databases[client]["database"] = connection[client]
-        databases[client]["capturescol"]  =  databases[client]["database"]["captures"]
-        databases[client]["capturescol"].drop() 
-        databases[client]["capturescol"]  =  databases[client]["database"]["captures"]
+        if client not in clients_to_ignore:
+            databases[client] = {}
+            databases[client]["database"] = connection[client]
+            databases[client]["capturescol"]  =  databases[client]["database"]["captures"]
+            databases[client]["capturescol"].drop() 
+            databases[client]["capturescol"]  =  databases[client]["database"]["captures"]
 
-        for room in data[client]:
-            for sensor_idx, sensor in enumerate(data[client][room], start=0):
-                for capture in data[client][room][sensor_idx]["generated_data"]:
-                    document = capture
-                    print(document)
-                    databases[client]["capturescol"].insert_one(document)
+            for room in data[client]:
+                for sensor_idx, sensor in enumerate(data[client][room], start=0):
+                    for capture in data[client][room][sensor_idx]["generated_data"]:
+                        document = capture
+                        print(document)
+                        databases[client]["capturescol"].insert_one(document)
